+++ THLabel/THLabel.h
@@ -1,16 +1,18 @@
 //
 //  THLabel.h
 //
-//  Version 1.2
+//  Version 1.3
 //
 //  Created by Tobias Hagemann on 11/25/12.
-//  Copyright (c) 2013 tobiha.de. All rights reserved.
+//  Copyright (c) 2014 tobiha.de. All rights reserved.
 //
 //  Original source and inspiration from:
 //  FXLabel by Nick Lockwood,
 //  https://github.com/nicklockwood/FXLabel
 //  KSLabel by Kai Schweiger,
 //  https://github.com/vigorouscoding/KSLabel
+//  GTMFadeTruncatingLabel by Google,
+//  https://code.google.com/p/google-toolbox-for-mac/source/browse/trunk/iPhone/
 //
 //  Big thanks to Jason Miller for showing me sample code of his implementation
 //  using Core Text! It inspired me to dig deeper and move away from drawing
@@ -42,11 +44,18 @@
 
 #import <UIKit/UIKit.h>
 
-typedef enum {
+typedef NS_ENUM(NSInteger, THLabelStrokePosition) {
 	THLabelStrokePositionOutside,
 	THLabelStrokePositionCenter,
 	THLabelStrokePositionInside
-} THLabelStrokePosition;
+};
+
+typedef NS_OPTIONS(NSUInteger, THLabelFadeTruncatingMode) {
+	THLabelFadeTruncatingModeNone = 0,
+	THLabelFadeTruncatingModeTail = 1 << 0,
+	THLabelFadeTruncatingModeHead = 1 << 1,
+	THLabelFadeTruncatingModeHeadAndTail = THLabelFadeTruncatingModeHead | THLabelFadeTruncatingModeTail
+};
 
 @interface THLabel : UILabel
 
@@ -65,6 +74,8 @@ typedef enum {
 @property (nonatomic, copy) NSArray *gradientColors;
 @property (nonatomic, assign) CGPoint gradientStartPoint;
 @property (nonatomic, assign) CGPoint gradientEndPoint;
+
+@property (nonatomic, assign) THLabelFadeTruncatingMode fadeTruncatingMode;
 
 @property (nonatomic, assign) UIEdgeInsets textInsets;